refactor time dimension to be backend agnostic
